Add a whitespace-nowrap to the alert-body.
This fixes #377

We added a whitespace nowrap to ensue the body of the dialog
is never wrapped, regardless of the size of the body.
Because this might push the body of the alert over the close-X, we add a
padding to the right, to ensure the X and the body never overlap.