Allow wrapping on long texts, but avoid wrapping of "ETH address saved!"
This is an addition to commit:37c6869. After feedback, the whitespace
nowrap was considered not adequate, hence we remove it.
The padding creates enough space to avoid wrapping longer texts, because
it reserves space for the close-button. When that space is not reserved,
the close button pushes the content-box and forces it narrower,
resulting in the wrapping.

This fixes #377